+++ renderer/src/modules/api/patcher.js
@@ -37,7 +37,7 @@ class Patcher {
 
     /**
      * This method patches onto another function, allowing your code to run instead. 
-     * Using this, you are also able to modify the return value, using the return of your code instead.
+     * Using this, you are able to replace the original completely. You can still call the original manually if needed.
      * @param {string} caller Name of the caller of the patch function.
      * @param {object} moduleToPatch Object with the function to be patched. Can also be an object's prototype.
      * @param {string} functionName Name of the function to be patched.
@@ -55,8 +55,8 @@ class Patcher {
     }
 
     /**
-     * This method patches onto another function, allowing your code to run instead. 
-     * Using this, you are also able to modify the return value, using the return of your code instead.
+     * This method patches onto another function, allowing your code to run afterwards. 
+     * Using this, you are able to modify the return value after the original method is run.
      * @param {string} caller Name of the caller of the patch function.
      * @param {object} moduleToPatch Object with the function to be patched. Can also be an object's prototype.
      * @param {string} functionName Name of the function to be patched.